Consolidate class attendance admin route
