+++ web/app/routes/manage/nav.ts
@@ -26,7 +26,6 @@ export const manageSections: ManageNavSection[] = [
     defaultCollapsed: false,
     items: [
       { to: '/manage/class-attendance', label: 'Class attendance', description: 'Attendance by class and student.' },
-      { to: '/manage/attendance', label: 'Attendance (raw)', description: 'Raw class_attendance table rows for direct debugging.' },
       { to: '/manage/workshop-enrollment', label: 'Workshop enrollments', description: 'Pending, waitlisted, approved, and rejected workshop enrollments.' },
       { to: '/manage/class', label: 'Classes', description: 'Individual class schedule entries.' },
       { to: '/manage/gift-cards', label: 'Gift cards', description: 'Upload and process gift card batches.' },
